update tagsinput ui component - done
