+++ src/crm/Accounts/EditAccount.js
@@ -16,8 +16,6 @@ import axios from 'axios';
 
 export default function EditAccount(props) {
 
-  // console.log(props);
-
   const [accountObject, setAccountObject] = useState({
     name: '', website: '', phone: '', email: '',
       billing_address_line: '', billing_street: '', billing_postcode: '',
@@ -31,9 +29,7 @@ export default function EditAccount(props) {
   const [tags, setTags] = useState([]);
   const [file, setFile] = useState([]);
   const [isValidations, setIsValidations] = useState('true');
-  const [errors, setErrors] = useState({});
-  const [tagErrorStyle, setTagErrorStyle] = useState(''); 
-  const [invalidTag, setIsInvalidTag] = useState([]);    
+  const [errors, setErrors] = useState({});  
 
   useEffect(() => {
     getContacts();
@@ -86,54 +82,23 @@ export default function EditAccount(props) {
     })    
   }
 
-  // console.log(accountObject);
-
   const handleChange = (e) => {    
     setAccountObject({...accountObject, [e.target.name]: e.target.value})    
   }  
-
-  const addTags = event => {        
-    event.preventDefault();
-    if (event.key === 'Enter' && event.target.value !== "") {       
-      let val = event.target.value;             
-      if(!tags.includes(val)) {        
-        setTags([...tags, event.target.value]);         
-        setIsInvalidTag('');                
-      }
-      event.target.value="";
-    }    
-  }
-
-  const handleTag = (e) => {    
-    e.preventDefault();
-    if(tags.includes(e.target.value)) {
-      setTagErrorStyle('invalid_tag');
-    }else{
-      setTagErrorStyle('');      
-    }    
-    setIsInvalidTag(e.target.value);    
-  }
-
-  const removeTags = index => {        
-    setTags([...tags.filter(tag => tags.indexOf(tag) !== index)]);
-  }
   
   const fileUpload = (e) => {       
     let filesArray = [...accountObject.files];  
     let newFile = e.target.files[0];
-    filesArray.push(newFile);
-    // console.log(newFile);    
+    filesArray.push(newFile);       
     setAccountObject({...accountObject, files: filesArray});
   }
     
 
   const removeFile = (createdOn) => {
       let dupFiles = [...accountObject.files];
-      let remainingFiles = dupFiles.filter(file => file.created_on !== createdOn);
-      console.log(remainingFiles);
+      let remainingFiles = dupFiles.filter(file => file.created_on !== createdOn);      
       setAccountObject({...accountObject, files: remainingFiles});
-  }
-  console.log(accountObject);
+  }  
 
   const updateAccount = (e) => {
     e.preventDefault();
@@ -244,14 +209,8 @@ export default function EditAccount(props) {
                         <ReactSelect elementSize="col-md-12" labelName="Users" isDisabled={true}/>
                         <ReactSelect elementSize="col-md-12" labelName="Assigned To"/>
                         <SelectComponent  elementSize="col-md-12" labelName="Status" attrName="status" attrPlaceholder="Status" attrId="id_status" 
-                                          value={accountObject.status} getInputValue={handleChange} options={twoStatus}/>                           
-                        <TagsInput tags={tags}
-                              removeTags={(index) => removeTags(index)}
-                              addTags={(event) => addTags(event)}
-                              value={invalidTag}
-                              handleTag={(e) => handleTag(e)}
-                              tagErrorStyle={tagErrorStyle}/>
-                       
+                                          value={accountObject.status} getInputValue={handleChange} options={twoStatus}/>                                                   
+                        <TagsInput type="edit" sendTags={tags} getTags={setTags}/>
                         <FileInput  elementSize="col-md-12" labelName="Attachment" attrName="account_attachment" inputId="id_file"  
                                     getFile={fileUpload}/>
                         <div>
